requirement.txt
Django==1.7.1
Markdown==2.6.1
PyJWT==1.1.0
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
astroid==1.3.6
billiard==3.3.0.20
bitly-api==0.3
boto==2.38.0
celery==3.1.18
cloudinary==1.0.21
coverage==3.7.1
django-celery==3.1.16
django-dotenv==1.3.0
django-encrypted-fields==1.1.2
django-extensions==1.4.9
django-password-reset==0.7
django-push-notifications==1.2.1
django-rest-swagger==0.2.9
django-storages==1.1.8
django-uuidfield==0.5.0
djangorestframework==3.0.1
ipdb==0.8
ipython==2.3.1
kombu==3.0.26
logilab-common==0.63.2
oauthlib==0.7.2
psycopg2==2.5.4
pyasn1==0.1.9
pycrypto==2.6.1
pylint==1.4.3
python-dateutil==1.5
python-keyczar==0.716
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
seacucumber==1.5.2
six==1.8.0
wsgiref==0.1.2
